refactor: use reducer for app state
