+++ packages/core/lib/use-puck-history.ts
@@ -1,6 +1,6 @@
-import type { Data } from "../types/Config";
+import type { AppData } from "../types/Config";
 import { useEffect } from "react";
-import { PuckAction } from "./reducer";
+import { PuckAction } from "../reducer";
 import { useActionHistory } from "./use-action-history";
 import { useHotkeys } from "react-hotkeys-hook";
 import EventEmitter from "event-emitter";
@@ -10,14 +10,14 @@ import { diff, applyChange, revertChange } from "deep-diff";
 const DEBOUNCE_TIME = 250;
 const RECORD_DIFF = "RECORD_DIFF";
 export const historyEmitter = EventEmitter();
-export const recordDiff = (newData: Data) =>
-  historyEmitter.emit(RECORD_DIFF, newData);
+export const recordDiff = (newAppData: AppData) =>
+  historyEmitter.emit(RECORD_DIFF, newAppData);
 
 export function usePuckHistory({
-  data,
+  appData,
   dispatch,
 }: {
-  data: Data;
+  appData: AppData;
   dispatch: (action: PuckAction) => void;
 }) {
   const { canForward, canRewind, rewind, forward, record } = useActionHistory();
@@ -26,10 +26,10 @@ export function usePuckHistory({
   useHotkeys("meta+shift+z", forward, { preventDefault: true });
   useHotkeys("meta+y", forward, { preventDefault: true });
 
-  const [snapshot] = useDebounce(data, DEBOUNCE_TIME);
+  const [snapshot] = useDebounce(appData, DEBOUNCE_TIME);
 
-  const handleRecordDiff = useDebouncedCallback((newData: Data) => {
-    const _diff = diff(snapshot, newData);
+  const handleRecordDiff = useDebouncedCallback((newAppData: AppData) => {
+    const _diff = diff(snapshot, newAppData);
 
     if (!_diff) {
       return;
@@ -37,20 +37,21 @@ export function usePuckHistory({
 
     record({
       forward: () => {
-        const target = structuredClone(data);
+        const target = structuredClone(appData);
         _diff.reduce((target, change) => {
           applyChange(target, true, change);
           return target;
         }, target);
-        dispatch({ type: "set", data: target });
+        dispatch({ type: "setData", data: target.data });
       },
       rewind: () => {
-        const target = structuredClone(data);
+        const target = structuredClone(appData);
         _diff.reduce((target, change) => {
           revertChange(target, true, change);
           return target;
         }, target);
-        dispatch({ type: "set", data: target });
+
+        dispatch({ type: "setData", data: target.data });
       },
     });
   }, DEBOUNCE_TIME);
